+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+
+## [UNRELEASED] - 2026-02-27
+
+### Added
+
+- Improved logging messages
+
+### Changed
+
+- Point feature indicator revisions
+- `forest` multi-polygon is initially hidden
+
+
 ## [0.1.0] - 2026-02-26
 
 Initial release